+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 
 - **Landingzones**
     - Send email when zone status is set as ``MOVED`` or ``FAILED`` (#280)
+    - Unit tests for ``ZoneStatusSetAPIView``
 - **Projectroles**
     - Helper ``email.send_generic_mail()`` (#280)
 - **Samplesheets**